cash_register/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
jinja2
python-multipart